DynamicDataset
Removes outdated 'DetectorDataset'. Renames old DynamicPipeline (which was a dataset class) as DynamicDataset. Moves DynamicDataset from its own 'pipeline' file to file with all other datasets. Removes most cropping functionality from DynamicDataset (to be implemented in pipeline), and changes inheritance to be from HeatmapDataset.

It should be entirely possible to eliminate DynamicDataset in the future, and just use HeatmapDataset, but this requires refactoring the entire codebase to accept different inputs in the dataset creation function in order to access the resize_dims info.