Create pipelines.py
Defines DynamicPipeline as a LigtningModule containing COM and POS sub-models for detection and tracking. Initial pass is not fully functional. Functioning elements are currently initialization and prepare_batch method, which takes a batch of labeled or unlabeled data, pushes it through COM net, gets crop parameters for each image, and then loops over images to crop and resize data for POS network. It also updates info in a new dictionary containing bbox of new images and transformed keypoints in the new image coordinates.